+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## [1.1.1]
+
+- fix: keep custom http client with converter [#100](https://github.com/supabase/postgrest-dart/pull/100)
+
+## [1.1.0]
+
+- fix: implement catchError [#97](https://github.com/supabase-community/postgrest-dart/pull/97)
+- feat: add generic types to `.select()` [#94](https://github.com/supabase-community/postgrest-dart/pull/94)
+  ```dart
+  // data is `List<Map<String, dynamic>>`
+  final data = await supabase.from<List<Map<String, dynamic>>>('users').select();
+
+  // data is `Map<String, dynamic>`
+  final data = await supabase.from<Map<String, dynamic>>('users').select().eq('id', myId).single();
+  ```
+
+## [1.0.1]
+
+- fix: calling `.select()` multiple times will override the previous `.select()` [#95](https://github.com/supabase-community/postgrest-dart/pull/95)
+
+## [1.0.0]
+
+- chore: v1.0.0 release 🚀
+- BREAKING: set minimum Dart SDK version to 2.15.0 [#92](https://github.com/supabase-community/postgrest-dart/pull/92)
+
 ## [1.0.0-dev.4]
 
 - fix: update insert documentation to reflect new `returning` behavior [#88](https://github.com/supabase-community/postgrest-dart/pull/88)